Remove "position" arguments from roles and channels.
Trying to manage positions using a global index-based system, when Terraform manages resources in arbitrary order on a resource-by-resource basis, does not work. There are also bugs in the implementation that can lead to orphaned resources in some cases if the reordering logic is not successful.

This can probably be implemented using a separate ordering resource, but for now we're going to handle ordering manually.